third_party/blink/web_tests/platform/generic/http/tests/images/document-policy-lossless-images-strict-max-bpp-expected.txt

CONSOLE ERROR: Document policy violation: Image bpp (byte per pixel) exceeds max value set in lossless-images-strict-max-bpp.
CONSOLE ERROR: Document policy violation: Image bpp (byte per pixel) exceeds max value set in lossless-images-strict-max-bpp.
